update jaxb model to parse both schema versions
